refactor: Phase 3 slice 3g — player/syncplay/debug routers, Mp4StatsService, collapse registerIpcHandlers (#84)
Final slice of the IPC-router refactor (epic #84, tracked in #102).
Migrates the last 22 handlers out of `registerIpcHandlers()` into three
per-domain routers, extracts the mp4-faststart sampling state into a
dedicated service, and collapses `registerIpcHandlers()` to a thin
deps-wiring wrapper around `registerIpcRouters({...})`.

3 new routers:
- player.ipc.ts   — 10 PLAYER_* handlers + the local `extractFirstSubtitle`
                    helper (only player handlers used it)
- syncplay.ipc.ts — 7 SYNCPLAY_* handlers; thin wrappers over the
                    `./syncplay` singleton
- debug.ipc.ts    — 5 quality-mismatch + mp4-stats debug channels; the
                    session-scoped `qualityMismatches` Map is closure
                    state inside the router

1 new service:
- services/mp4-stats — `createMp4StatsService({ store })` owns the
                       per-session probed-paths Set + the read-modify-
                       write serialization chain. Consumed by the
                       `downloadManager.onVideoDownloaded` hook in
                       `index.ts`, `player:find-local-file`, and the
                       `DEBUG_GET/RESET_MP4_STATS` handlers.

`AppDeps` gains `streamingService` + `mp4StatsService`. The
`registerIpcRouters` doc-comment is updated to reflect 3g being terminal.

`index.ts` cleanup:
- `registerIpcHandlers()` shrinks from ~600 lines to ~30 lines (just the
  deps wiring).
- Dropped now-dead imports: `ipcMain`, `Ffmpeg`, `randomUUID`, `CHANNELS`,
  `ChildProcess`, `MseSession`, `MseOpenResult`, `probeMp4Faststart`,
  `SyncplayConfig`, `SyncplayFileInfo`.
- Stale Phase-4 cache-GC comment removed (its referenced helpers were
  folded into `SkipAnalysisService` in earlier phases).
- Net ~640 LOC removed.

Behavior-identical. `STORAGE_MOVE_TO_COLD`-style broadcasts and
`recordMp4FaststartCheck` semantics preserved. `SYNCPLAY_CONNECT`'s
spread of `store.get('syncplay')` needed a `Record<string, unknown>`
cast in the router context (the generic `StorageService.get` returns
`unknown`); all 5 fields are still overridden so behavior is identical.

DESIGN.md: index.ts row retargeted to the final-state description,
three router rows added, `mp4-stats` service row added, the streaming
row's trailing sentence retargeted to `player.ipc.ts`.
